renamed sensorsconditioners as samplesconditioners, added logger and channels_loggers
