+++ ardas/logger.py
@@ -3,37 +3,37 @@ from time import sleep
 from ardas.sampler import W1Sampler
 from ardas.sensor_tools import TempSensor, generate_w1temp_sensors
 from ardas.samples_conditioners import generate_w1temp_sensor_samples_conditioners
+from ardas.conditioner import Conditioner
 import queue
 
 
-class Conditioner(Thread):
-    def __init__(self, stop_event, samples_conditioners, sampler_queue, logger_queue):
+class Logger(Thread):
+    def __init__(self, stop_event, channels_loggers, logger_queue, loggers=None):
         Thread.__init__(self)
         self.stop_event = stop_event
-        self.__samples_conditioners = samples_conditioners
-        self.sampler_queue = sampler_queue
+        self.__channels_loggers = channels_loggers
         self.logger_queue = logger_queue
+        self.loggers = loggers
 
     @property
-    def samples_conditioners(self):
-        """ Gets and sets sensor id
+    def channels_loggers(self):
+        """ Gets and sets channels loggers
         """
-        return self.__samples_conditioners
+        return self.__channels_loggers
 
-    @samples_conditioners.setter
-    def samples_conditioners(self, val):
-        self.__samples_conditioners = val
+    @channels_loggers.setter
+    def channels_loggers(self, val):
+        self.__channels_loggers = val
 
     def run(self):
         while not self.stop_event.isSet():
             try:
-                sample = self.sampler_queue.get(timeout=1.0)
-                if sample is not None:
-                    s_id = sample['tags']['sensor']
-                    sc = self.samples_conditioners[s_id]
-                    data = sc.output(sample)
-                    data['tags']['channel'] = '%s' % sc.channel_name
-                    self.logger_queue.put(data)
+                record = self.logger_queue.get(timeout=1.0)
+                if record is not None:
+                    print(record)
+                    ch_id = record['tags']['channel']
+                    cl = self.channels_loggers[ch_id]
+                    cl.log()
             except queue.Empty:
                 pass
 
@@ -48,13 +48,16 @@ if __name__ == '__main__':
         sensors = generate_w1temp_sensors(7)
     stop = Event()
     sampler = W1Sampler(stop_event=stop, interval=5, sensors=sensors, sampler_queue=sampler_queue)
-    samples_conditioners = generate_w1temp_sensor_samples_conditioners(sensors=sensors)
-    conditioner = Conditioner(stop_event=stop, samples_conditioners=samples_conditioners, sampler_queue=sampler_queue,
+    sensors_conditioners = generate_w1temp_sensor_samples_conditioners(sensors=sensors)
+    channels_loggers = None
+    conditioner = Conditioner(stop_event=stop, samples_conditioners=sensors_conditioners, sampler_queue=sampler_queue,
                               logger_queue=logger_queue)
+    logger = Logger(stop_event=stop, channels_loggers=channels_loggers, logger_queue=logger_queue)
     sampler.start()
     conditioner.start()
+    logger.start()
     k = 0
-    kmax = 20
+    kmax = 10
     while k < kmax:
         k += 1
         sleep(1.0)
