+++ ardas/sensor_tools.py
@@ -8,7 +8,7 @@ try:
 
     class TempSensor(W1ThermSensor):
         def __init__(self, sensor_type=None, sensor_id=None, name=None):
-            super(TempSensor, self).__init__(sensor_type=sensor_type, sensor_id=sensor_id)
+            super(TempSensor, self).__init__(sensor_type=sensor_type, sensor_id=sensor_id)  # TODO: check this use of super...
             self.__name = name
 
         @property
@@ -38,7 +38,7 @@ except:
 
     class TempSensor(FakeTempSensor):
         def __init__(self):
-            super(TempSensor, self).__init__()
+            super(TempSensor, self).__init__()   # TODO: check this use of super...
 
 cur_dir = Path(__file__).resolve().parent  # os.path.dirname(os.path.realpath(__file__))
 
@@ -55,174 +55,6 @@ def load_sensor(id):
     return sensor
 
 
-def polynomial(value, coefs):
-    """Compute polynomial using Horner method
-
-    :param value: given value of the variable
-    :param coefs: coefficients of the polynomial
-    :return: evaluation of polynomial for the given value of the variable
-    :rtype: float
-    """
-
-    result = coefs[-1]
-    for i in range(-2, -len(coefs) - 1, -1):
-        result = result * value + coefs[i]
-    assert isinstance(result, float)
-    return result
-
-
-def running_average(n):  # TODO: add self ?
-    """Compute a running average (not centered)
-
-    :param n: number of former samples used to compute the running average
-    :return: evaluation of polynomial for the given value of the variable (freq)
-    :rtype: float
-    """
-
-    l = []
-    average = None
-    while True:
-        new_elt = yield average
-        if len(l) == n:
-            del l[0]
-        l.append(new_elt)
-        if len(l) > 0:
-            average = sum(l) / len(l)
-        else:
-            average = np.nan
-
-
-def no_processing(self, sample, params=None):
-    """Simple copy of the value
-
-    :param self: calling object
-    :param sample: value from sensor
-    :param params: None
-    :return: same value
-    """
-
-    assert (type(sample) is dict), 'sample should be a dict'
-    sample['fields']['format'] = self.output_format
-    sample['fields']['units'] = self.units
-    sample['fields']['quantity'] = self.quantity
-    sample['tags']['processing'] = 'no processing'
-    return sample
-
-
-class SensorConditioner(object):
-    """ A class to handle sensors
-    """
-    def __init__(self, sensor=None, channel_name='', processing_method=None, processing_parameters=None,
-                 quantity='', units='', output_format='%11.4f', log_output=True):
-        self.sensor = sensor
-        self.__channel_name = channel_name
-        self.processing_method = processing_method
-        self.processing_parameters = processing_parameters
-        self.__quantity = quantity
-        self.__units = units
-        self.__output_format = output_format
-        self.__log = log_output
-
-    @property
-    def channel_name(self):
-        """Gets and sets sensor name
-        """
-        return self.__channel_name
-
-    @channel_name.setter
-    def channel_name(self, val):
-        self.__channel_name = str(val)
-
-    @property
-    def quantity(self):
-        """Gets and sets sensor units
-        """
-        return self.__quantity
-
-    @quantity.setter
-    def quantity(self, val):
-        self.__quantity = str(val)
-
-    @property
-    def units(self):
-        """Gets and sets sensor units
-        """
-        return self.__units
-
-    @units.setter
-    def units(self, val):
-        self.__units = val
-
-    @property
-    def output_format(self):
-        """Gets and sets sensor units
-        """
-        return self.__output_format
-
-    @output_format.setter
-    def output_format(self, val):
-        self.__output_format = val
-
-    def output(self, sample):
-        """Outputs an output computed using the processing method and parameters
-
-        input sample should be in the form {time: value}
-
-        :return: processed quantity
-        :rtype: float
-        """
-
-        output = self.processing_method(self, sample, self.processing_parameters)
-        return output
-
-    def output_repr(self, value):
-        """Gets a representation of the output
-
-        :return: representation of the processed quantity
-        :rtype: string
-        """
-
-        try:
-            s = self.output_format + ' ' + self.units
-            calibrated_output = s % self.output(value)
-        except Exception as e:
-            calibrated_output = '*** error : %s ***' % e
-        assert isinstance(calibrated_output, str)
-        return calibrated_output
-
-    def save(self):  # TODO deal with sensors sensor conditioner and channel
-        """Save the sensor as a serialized object to a file """
-        f_name = cur_dir / 'sensor_' + self.name + '.ssr'
-        if Path(f_name).exists():
-            logging.warning('Sensor file ' + f_name + ' already exists, unable to save sensor')
-        else:
-            with open(f_name, 'wb') as sensor_file:
-                dump(self, sensor_file)
-
-
-class FMSensorConditioner(SensorConditioner):
-    """A subclass of the Sensor object with a simpler interface"""
-    def __init__(self, channel_name='0000', processing_method=polynomial, processing_parameters=(0., 1., 0., 0., 0.),
-                 quantity='freq.', units='Hz'):
-        SensorConditioner.__init__(channel_name=channel_name, processing_method=processing_method,
-                                   processing_parameters=processing_parameters, quantity=quantity, units=units)
-
-
-class UncalibratedFMSensorConditioner(FMSensorConditioner):
-    """A subclass of the FMsensor object with a simpler interface"""
-    def __init__(self, channel_name='0000', log_output=True):
-        FMSensorConditioner.__init__(channel_name=channel_name, log_output=log_output)
-
-
-class W1TempSensorConditioner(SensorConditioner):
-    def __init__(self, sensor, channel_name, processing_method=no_processing, processing_parameters=None):
-        SensorConditioner.__init__(self, sensor=sensor, channel_name=channel_name, processing_method=processing_method,
-                                   processing_parameters=processing_parameters, log_output=True)
-        self.quantity = 'temp.'
-        self.units = '°C'
-        self.output_format = '%6.3f'
-
-
 def generate_w1temp_sensors(nb_sensor=2):
     sensors = []
     for i in range(nb_sensor):
@@ -232,20 +64,5 @@ def generate_w1temp_sensors(nb_sensor=2):
     return sensors
 
 
-def generate_w1temp_sensors_conditioners(nb_sensor=2, sensors=None):
-    if sensors is None:
-        sensors = generate_w1temp_sensors(nb_sensor)
-    else:
-        nb_sensor = len(sensors)
-    sensors_conditioners = {}
-    for i in range(nb_sensor):
-        s = W1TempSensorConditioner(sensor=sensors[i], channel_name='T%03d' % i)
-        sensors_conditioners.update({'%s%s' % (sensors[i].slave_prefix, sensors[i].id): s})
-    return sensors_conditioners
-
-
 if __name__ == '__main__':
-    t = polynomial(25000, [-16.9224032438, 0.0041525221, -1.31475837290789e-07, 2.39122208189129e-12,
-                           -1.72530800355418e-17])
-    print(t)
-    print(cur_dir)
+    pass
